pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alteris-listener"
version = "0.2.0"
description = "Query local Mac data sources (Mail, iMessage, Slack, Calendar, Granola) with LLMs"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "requests>=2.28",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
gemini = ["google-genai>=1.0"]
claude = ["anthropic>=0.40"]
slack = ["slack-sdk>=3.27"]
all = [
    "google-genai>=1.0",
    "anthropic>=0.40",
    "slack-sdk>=3.27",
]

[project.scripts]
alteris-listener = "alteris_listener.main:cli"

[tool.setuptools.packages.find]
include = ["alteris_listener*"]
